reality-engine/requirements.txt
feedparser==6.0.10
newspaper3k==0.2.8
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
httpx==0.25.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
spacy==3.7.2
vaderSentiment==3.3.2
langid==1.1.6
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
